move route description from Qt Router to new class RouteDescriptionBuilder
