+++ libosmscout-client-qt/include/osmscout/Router.h
@@ -21,10 +21,7 @@
 #ifndef ROUTER_H
 #define ROUTER_H
 
-#include <memory>
-
-#include <QObject>
-#include <QSettings>
+#include <osmscout/RouteDescriptionBuilder.h>
 
 #include <osmscout/DataTileCache.h>
 #include <osmscout/DBThread.h>
@@ -35,74 +32,12 @@
 
 #include <osmscout/private/ClientQtImportExport.h>
 
+#include <memory>
+
+#include <QObject>
+#include <QSettings>
+
 Q_DECLARE_METATYPE(osmscout::Vehicle)
-
-/**
- * \ingroup QtAPI
- */
-class OSMSCOUT_CLIENT_QT_API RouteStep : public QObject
-{
-    Q_OBJECT
-    Q_PROPERTY(QString distance READ getDistance)
-    Q_PROPERTY(QString distanceDelta READ getDistanceDelta)
-    Q_PROPERTY(QString time READ getTime)
-    Q_PROPERTY(QString timeDelta READ getTimeDelta)
-    Q_PROPERTY(QString description READ getDescription)
-
-public:
-  QString distance;
-  QString distanceDelta;
-  QString time;
-  QString timeDelta;
-  QString description;
-
-public:
-  RouteStep();
-  RouteStep(const RouteStep& other);
-
-  RouteStep& operator=(const RouteStep& other);
-
-  QString getDistance() const
-  {
-      return distance;
-  }
-
-  QString getDistanceDelta() const
-  {
-      return distanceDelta;
-  }
-
-  QString getTime() const
-  {
-      return time;
-  }
-
-  QString getTimeDelta() const
-  {
-      return timeDelta;
-  }
-
-  QString getDescription() const
-  {
-      return description;
-  }
-};
-
-/**
- * \ingroup QtAPI
- */
-struct RouteSelection
-{
-  osmscout::RouteData        routeData;
-  osmscout::RouteDescription routeDescription;
-  QList<RouteStep>           routeSteps;
-  osmscout::Way              routeWay;
-};
-
-typedef std::shared_ptr<RouteSelection> RouteSelectionRef;
-
-Q_DECLARE_METATYPE(RouteSelection)
-Q_DECLARE_METATYPE(RouteSelectionRef)
 
 /**
  * \ingroup QtAPI
@@ -196,44 +131,6 @@ signals:
 
 private:
   void GetCarSpeedTable(std::map<std::string,double>& map);
-
-  void DumpStartDescription(RouteSelection &route,
-                            const osmscout::RouteDescription::StartDescriptionRef& startDescription,
-                            const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-
-  void DumpTargetDescription(RouteSelection &route, 
-                             const osmscout::RouteDescription::TargetDescriptionRef& targetDescription);
-
-  void DumpTurnDescription(RouteSelection &route,
-                           const osmscout::RouteDescription::TurnDescriptionRef& turnDescription,
-                           const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription,
-                           const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
-                           const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-
-  void DumpRoundaboutEnterDescription(RouteSelection &route,
-                                      const osmscout::RouteDescription::RoundaboutEnterDescriptionRef& roundaboutEnterDescription,
-                                      const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
-
-  void DumpRoundaboutLeaveDescription(RouteSelection &route,
-                                      const osmscout::RouteDescription::RoundaboutLeaveDescriptionRef& roundaboutLeaveDescription,
-                                      const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-
-  void DumpMotorwayEnterDescription(RouteSelection &route,
-                                    const osmscout::RouteDescription::MotorwayEnterDescriptionRef& motorwayEnterDescription,
-                                    const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
-
-  void DumpMotorwayChangeDescription(RouteSelection &route,
-                                     const osmscout::RouteDescription::MotorwayChangeDescriptionRef& motorwayChangeDescription);
-  
-  void DumpMotorwayLeaveDescription(RouteSelection &route,
-                                    const osmscout::RouteDescription::MotorwayLeaveDescriptionRef& motorwayLeaveDescription,
-                                    const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
-                                    const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-  
-  void DumpNameChangedDescription(RouteSelection &route,
-                                  const osmscout::RouteDescription::NameChangedDescriptionRef& nameChangedDescription);
-
-  void GenerateRouteSteps(RouteSelection &route);
 
   void ProcessRouteRequest(osmscout::MultiDBRoutingServiceRef &routingService,
                            const LocationEntryRef &start,
